Move issue context class to separate file. Cleanup detector class
