+++ server/bundled/tools/python/nplusone/nplusone_analyzer.py
@@ -1,53 +1,13 @@
 import ast
 import uuid
 
-from typing import List, Dict, Any, Set
+from typing import List, Dict, Any
 
 from log import LOGGER
 from constants import QUERY_METHODS
 
 from nplusone.scorer import NPlusOneScorer
-
-class GlobalContext:
-    def __init__(self):
-        self.optimized_querysets: Set[ast.AST] = set()
-        self.optimized_variables: Dict[str, ast.AST] = {}
-        self.variable_assignments: Dict[str, ast.AST] = {}
-
-    def add_optimized_queryset(self, queryset_node: ast.AST):
-        LOGGER.debug(f"Adding optimized queryset: {queryset_node}")
-        self.optimized_querysets.add(queryset_node)
-        LOGGER.debug(f"Optimized querysets: {self.optimized_querysets}")
-
-    def add_optimized_variable(self, var_node: ast.AST, queryset_node: ast.AST):
-        LOGGER.debug(f"Tracking variable: {var_node.id} -> {queryset_node}")
-        self.optimized_variables[var_node.id] = queryset_node
-        LOGGER.debug(f"Optimized variables: {self.optimized_variables}")
-
-    def is_queryset_optimized(self, queryset_node: ast.AST) -> bool:
-        LOGGER.debug(f"Checking if queryset is optimized: {queryset_node}")
-        LOGGER.debug(f"Current optimized querysets: {self.optimized_querysets}")
-        return queryset_node in self.optimized_querysets
-    
-    def is_variable_optimized(self, var_node: ast.AST) -> bool:
-        is_optimized = var_node in self.optimized_variables
-        LOGGER.debug(f"Checking if variable is optimized: {var_node.value} {is_optimized}")
-        return is_optimized
-
-    def get_queryset_for_variable(self, var_node: ast.AST) -> ast.AST:
-        LOGGER.debug(f"Trying to get queryset for variable: {var_node} {type(var_node)} {self.optimized_variables}")
-        return self.optimized_variables.get(var_node.id)
-    
-    def add_variable_assignment(self, var_node: ast.Name, assigned_value: ast.AST):
-        LOGGER.debug(f"Tracking variable assignment: {var_node.id} -> {assigned_value}")
-        self.variable_assignments[var_node.id] = assigned_value
-        LOGGER.debug(f"Variable assignments: {self.variable_assignments}")
-
-    def get_variable_assignment(self, var_node: ast.AST) -> ast.AST:
-        if isinstance(var_node, ast.Name):
-            return self.variable_assignments.get(var_node.id)
-        LOGGER.debug(f"Variable assignment lookup skipped for non-Name node: {ast.dump(var_node)}")
-        return None
+from nplusone.optimized_context import NPlusOneGlobalContext
 
 
 class ParentAwareNodeVisitor(ast.NodeVisitor):
@@ -69,7 +29,7 @@ class NPlusOneDetector:
         self.source_code = source_code
         self.model_cache = model_cache
         self.issues: List[Dict[str, Any]] = []
-        self.global_context = GlobalContext()
+        self.global_context = NPlusOneGlobalContext()
         self.detected_chains_global = set()
 
         LOGGER.info(f"Initializing N+1 detector with models: {model_cache}")
@@ -154,7 +114,6 @@ class NPlusOneDetector:
     def analyze_loop(self, loop: ast.For):
         if isinstance(loop.target, ast.Name) and isinstance(loop.iter, ast.Name):
             self.process_loop_target_and_iterable(loop.target, loop.iter)
-
         self.process_loop_nodes(loop)
 
     def track_variable_assignment(self, node: ast.AST, queryset_name: str):
@@ -386,17 +345,6 @@ class NPlusOneDetector:
         }
         LOGGER.debug(f"Adding issue: {issue}")
         self.issues.append(issue)
-
-    def get_enclosing_function_name(self, node: ast.AST) -> str:
-        """
-        Get the name of the function enclosing the node, if available.
-        """
-        current_node = node
-        while current_node:
-            if isinstance(current_node, ast.FunctionDef):
-                return current_node.name
-            current_node = getattr(current_node, 'parent', None)
-        return "Unknown"
 
     def get_problematic_code(self, node: ast.AST) -> str:
         """
